move BIND specific fns from ./index to lib/zonefile
