+++ test/index.js
@@ -3,58 +3,58 @@ const assert = require('assert')
 const fs     = require('fs')
 
 const RR = require('dns-resource-record')
-const zv = require('../index')
+const zf = require('../lib/zonefile')
 
 beforeEach(function () {
-  zv.zoneOpts = {}
+  zf.zoneOpts = {}
 })
 
 describe('parseZoneFile', function () {
 
   it('parses a blank line', async () => {
-    const r = await zv.parseZoneFile(`\n`)
+    const r = await zf.parseZoneFile(`\n`)
     // console.dir(r[0], { depth: null })
     assert.deepStrictEqual(r, [ '\n' ])
   })
 
   it('parses two blank lines', async () => {
-    const r = await zv.parseZoneFile(`\n\n`)
+    const r = await zf.parseZoneFile(`\n\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r, [ '\n', '\n' ])
   })
 
   it('parses a line with only whitespace', async () => {
-    const r = await zv.parseZoneFile(` \t\n`)
+    const r = await zf.parseZoneFile(` \t\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r, [ ' \t\n' ])
   })
 
   it('parses a comment line', async () => {
-    const r = await zv.parseZoneFile(`; blank comment\n`)
+    const r = await zf.parseZoneFile(`; blank comment\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r, [ '; blank comment\n' ])
   })
 
   it('parses a comment line with leading ws', async () => {
-    const r = await zv.parseZoneFile(` ; blank comment with leading ws\n`)
+    const r = await zf.parseZoneFile(` ; blank comment with leading ws\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r, [ '; blank comment with leading ws\n' ])
   })
 
   it('parses a $TTL line', async () => {
-    const r = await zv.parseZoneFile(`$TTL 86400\n`)
+    const r = await zf.parseZoneFile(`$TTL 86400\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r[0], { $TTL: 86400 })
   })
 
   it('parses a $TTL line with a comment', async () => {
-    const r = await zv.parseZoneFile(`$TTL 86400; yikers\n`)
+    const r = await zf.parseZoneFile(`$TTL 86400; yikers\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r[0], { $TTL: 86400, comment: '; yikers' })
   })
 
   it(`parses a SOA`, async () => {
-    const r = await zv.parseZoneFile(`example.com.   86400   IN  SOA ns1.example.com.    hostmaster.example.com. (
+    const r = await zf.parseZoneFile(`example.com.   86400   IN  SOA ns1.example.com.    hostmaster.example.com. (
                     2021102100    ; serial
                     16384   ; refresh
                     2048     ; retry
@@ -86,7 +86,7 @@ describe('parseZoneFile', function () {
   })
 
   it('parses a NS line', async () => {
-    const r = await zv.parseZoneFile(`cadillac.net.   14400   IN  NS  ns1.cadillac.net.\n`)
+    const r = await zf.parseZoneFile(`cadillac.net.   14400   IN  NS  ns1.cadillac.net.\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r[0], {
       name : 'cadillac.net.',
@@ -98,7 +98,7 @@ describe('parseZoneFile', function () {
   })
 
   it('parses an A line', async () => {
-    const r = await zv.parseZoneFile(`cadillac.net.   86400   IN  A   66.128.51.173\n`)
+    const r = await zf.parseZoneFile(`cadillac.net.   86400   IN  A   66.128.51.173\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r[0], {
       name   : 'cadillac.net.',
@@ -134,13 +134,13 @@ describe('parseZoneFile', function () {
 
   for (const t of testCAAs) {
     it(`parses CAA record: ${t.result.name}`, async () => {
-      const r = await zv.parseZoneFile(t.bind)
+      const r = await zf.parseZoneFile(t.bind)
       assert.deepStrictEqual(r[0], t.result)
     })
   }
 
   it('parses a CNAME line, absolute', async () => {
-    const r = await zv.parseZoneFile(`www 28800 IN  CNAME vhost0.theartfarm.com.\n`)
+    const r = await zf.parseZoneFile(`www 28800 IN  CNAME vhost0.theartfarm.com.\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r[0], {
       name : 'www',
@@ -152,8 +152,8 @@ describe('parseZoneFile', function () {
   })
 
   it('parses a CNAME line, relative', async () => {
-    zv.zoneOpts = { origin: 'theartfarm.com' }
-    const r = await zv.parseZoneFile(`www 28800 IN  CNAME vhost0\n`).then(zv.expandShortcuts)
+    zf.zoneOpts = { origin: 'theartfarm.com' }
+    const r = await zf.parseZoneFile(`www 28800 IN  CNAME vhost0\n`).then(zf.expandShortcuts)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r[0], new RR.CNAME({
       name : 'www.theartfarm.com.',
@@ -165,7 +165,7 @@ describe('parseZoneFile', function () {
   })
 
   it('parses a DNAME line', async () => {
-    const r = await zv.parseZoneFile(`_tcp 86400 IN  DNAME _tcp.theartfarm.com.\n`)
+    const r = await zf.parseZoneFile(`_tcp 86400 IN  DNAME _tcp.theartfarm.com.\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r[0], {
       name  : '_tcp',
@@ -177,7 +177,7 @@ describe('parseZoneFile', function () {
   })
 
   it('parses a DNSKEY record', async () => {
-    const r = await zv.parseZoneFile(
+    const r = await zf.parseZoneFile(
       `example.com. 86400 IN DNSKEY 256 3 5 ( AQPSKmynfzW4kyBv015MUG2DeIQ3
                                           Cbl+BBZH4b/0PY1kxkmvHjcZc8no
                                           kfzj31GajIQKY+5CptLr3buXA10h
@@ -199,7 +199,7 @@ describe('parseZoneFile', function () {
   })
 
   it('parses a DS record', async () => {
-    const r = await zv.parseZoneFile(
+    const r = await zf.parseZoneFile(
       `dskey.example.com. 86400 IN DS 60485 5 1 ( 2BB183AF5F22588179A53B0A
                                               98631FAD1A292118 )\n`)
     // console.dir(r, { depth: null })
@@ -216,7 +216,7 @@ describe('parseZoneFile', function () {
   })
 
   it('parses a HINFO line', async () => {
-    const r = await zv.parseZoneFile(`SRI-NIC.ARPA. HINFO   DEC-2060 TOPS20\n`)
+    const r = await zf.parseZoneFile(`SRI-NIC.ARPA. HINFO   DEC-2060 TOPS20\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r[0], {
       name: 'SRI-NIC.ARPA.',
@@ -228,7 +228,7 @@ describe('parseZoneFile', function () {
   })
 
   it('parses a LOC line', async () => {
-    const r = await zv.parseZoneFile(`rwy04l.logan-airport.boston. 3600 IN LOC 42 21 28.764 N 71 0 51.617 W -44m 2000m\n`)
+    const r = await zf.parseZoneFile(`rwy04l.logan-airport.boston. 3600 IN LOC 42 21 28.764 N 71 0 51.617 W -44m 2000m\n`)
     // console.dir(r, { depth: null })
     assert.deepStrictEqual(r[0], {
       name    : 'rwy04l.logan-airport.boston.',
@@ -257,7 +257,7 @@ describe('parseZoneFile', function () {
   })
 
   it('parses a MX line', async () => {
-    const r = await zv.parseZoneFile(`test.example.com. 3600 IN MX 0  mail.example.com.\n`)
+    const r = await zf.parseZoneFile(`test.example.com. 3600 IN MX 0  mail.example.com.\n`)
     assert.deepStrictEqual(r[0], {
       class     : 'IN',
       exchange  : 'mail.example.com.',
@@ -271,7 +271,7 @@ describe('parseZoneFile', function () {
   // cid.urn.arpa.   86400    IN    NAPTR 100    10    ""    ""    "!^urn:cid:.+@([^\\.]+\\.)(.*)$!\x02!i"   .
 
   it('parses a NS line', async () => {
-    const r = await zv.parseZoneFile(`example.com.  3600  IN  NS  ns1.example.com.\n`)
+    const r = await zf.parseZoneFile(`example.com.  3600  IN  NS  ns1.example.com.\n`)
     assert.deepStrictEqual(r[0], {
       class: 'IN',
       dname: 'ns1.example.com.',
@@ -282,7 +282,7 @@ describe('parseZoneFile', function () {
   })
 
   it('parses a PTR line', async () => {
-    const r = await zv.parseZoneFile(`2.2.0.192.in-addr.arpa. 86400  IN  PTR dhcp.example.com.\n`)
+    const r = await zf.parseZoneFile(`2.2.0.192.in-addr.arpa. 86400  IN  PTR dhcp.example.com.\n`)
     assert.deepStrictEqual(r[0], {
       class: 'IN',
       dname: 'dhcp.example.com.',
@@ -293,8 +293,8 @@ describe('parseZoneFile', function () {
   })
 
   it('parses a SOA line', async () => {
-    zv.zoneOpts = { origin: 'example.com', ttl: 3600 }
-    const r = await zv.parseZoneFile(`example.com.  IN  SOA ns1.example.com. matt.example.com. (
+    zf.zoneOpts = { origin: 'example.com', ttl: 3600 }
+    const r = await zf.parseZoneFile(`example.com.  IN  SOA ns1.example.com. matt.example.com. (
     1
     7200
     3600
@@ -323,7 +323,7 @@ describe('parseZoneFile', function () {
   })
 
   it('parses a TXT line', async () => {
-    const r = await zv.parseZoneFile(`oct2021._domainkey.example.com. 86400  IN  TXT "v=DKIM1;p=MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAoyUzGOTSOmakY8BcxXgi0mN/nFegLBPs7aaGQUtjHfa8yUrt9T2j6GSXgdjLuG3R43WjePQv3RHzc+bwwOkdw0XDOXiztn5mhrlaflbVr5PMSTrv64/cpFQKLtgQx8Vgqp7Dh3jw13rLomRTqJFgMrMHdhIibZEa69gtuAfDqoeXo6QDSGk5JuBAeRHEH27FriHulg5ob" "4F4lmh7fMFVsDGkQEF6jaIVYqvRjDyyQed3R3aTJX3fpb3QrtRqvfn/LAf+3kzW58AjsERpsNCSTD2RquxbnyoR/1wdGKb8cUlD/EXvqtvpVnOzHeSeMEqex3kQI8HOGsEehWZlKd+GqwIDAQAB"\n`)
+    const r = await zf.parseZoneFile(`oct2021._domainkey.example.com. 86400  IN  TXT "v=DKIM1;p=MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAoyUzGOTSOmakY8BcxXgi0mN/nFegLBPs7aaGQUtjHfa8yUrt9T2j6GSXgdjLuG3R43WjePQv3RHzc+bwwOkdw0XDOXiztn5mhrlaflbVr5PMSTrv64/cpFQKLtgQx8Vgqp7Dh3jw13rLomRTqJFgMrMHdhIibZEa69gtuAfDqoeXo6QDSGk5JuBAeRHEH27FriHulg5ob" "4F4lmh7fMFVsDGkQEF6jaIVYqvRjDyyQed3R3aTJX3fpb3QrtRqvfn/LAf+3kzW58AjsERpsNCSTD2RquxbnyoR/1wdGKb8cUlD/EXvqtvpVnOzHeSeMEqex3kQI8HOGsEehWZlKd+GqwIDAQAB"\n`)
     assert.deepStrictEqual(r[0], {
       name : 'oct2021._domainkey.example.com.',
       ttl  : 86400,
@@ -347,7 +347,7 @@ describe('parseZoneFile', function () {
     fs.readFile(file, (err, buf) => {
       if (err) throw err
 
-      zv.parseZoneFile(buf.toString()).then(r => {
+      zf.parseZoneFile(buf.toString()).then(r => {
         // console.dir(r, { depth: null })
         assert.equal(r.length, 41)
       })
@@ -359,7 +359,7 @@ describe('parseZoneFile', function () {
     fs.readFile(file, (err, buf) => {
       if (err) throw err
 
-      zv.parseZoneFile(buf.toString()).then(zv.expandShortcuts).then(r => {
+      zf.parseZoneFile(buf.toString()).then(zf.expandShortcuts).then(r => {
         // console.dir(r, { depth: null })
         assert.equal(r.length, 11)
       })
@@ -371,8 +371,8 @@ describe('parseZoneFile', function () {
     fs.readFile(file, (err, buf) => {
       if (err) throw err
 
-      zv.parseZoneFile(buf.toString())
-        .then(zv.expandShortcuts)
+      zf.parseZoneFile(buf.toString())
+        .then(zf.expandShortcuts)
         .then(r => {
           // console.dir(r, { depth: null })
           assert.equal(r.length, 15)
@@ -394,7 +394,7 @@ describe('expandShortcuts', function () {
   it('expands @ name to $ORIGIN', async () => {
     const input = JSON.parse(JSON.stringify(testCase))
     input[2].name = '@'
-    const out = await zv.expandShortcuts(input)
+    const out = await zf.expandShortcuts(input)
     assert.deepEqual(out, [ new RR.A({
       address: '1.2.3.4',
       class  : 'IN',
@@ -408,7 +408,7 @@ describe('expandShortcuts', function () {
     const input = JSON.parse(JSON.stringify(testCase))
     input[2].name = ''
 
-    const out = await zv.expandShortcuts(input)
+    const out = await zf.expandShortcuts(input)
     assert.deepEqual(out, [ new RR.A({
       address: '1.2.3.4',
       class  : 'IN',
@@ -435,7 +435,7 @@ describe('expandShortcuts', function () {
       type   : 'A',
     }
 
-    const out = await zv.expandShortcuts(input)
+    const out = await zf.expandShortcuts(input)
     assert.deepEqual(out[1], new RR.A({
       address: '1.2.3.4',
       class  : 'IN',
@@ -449,8 +449,8 @@ describe('expandShortcuts', function () {
     let input = JSON.parse(JSON.stringify(testCase))
     input = input.filter(e => !e.$TTL)
 
-    const r = await zv.parseZoneFile(`@ 55 IN  SOA ns1.cadillac.net. hostmaster.cadillac.net. (2021102100 16384 2048 604800 2560)\n`)
-    const out = await zv.expandShortcuts([ input[0], r[0], input[1] ])
+    const r = await zf.parseZoneFile(`@ 55 IN  SOA ns1.cadillac.net. hostmaster.cadillac.net. (2021102100 16384 2048 604800 2560)\n`)
+    const out = await zf.expandShortcuts([ input[0], r[0], input[1] ])
     assert.deepEqual(out[1], new RR.A({
       address: '1.2.3.4',
       class  : 'IN',
